requirements.txt
apispec==6.10.0
blinker==1.9.0
cachetools==6.2.0
click==8.3.1
cryptography==46.0.1
Flask==3.1.3
flask-smorest==0.47.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.2
packaging==26.0
psycopg[binary]
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
SQLAlchemy==2.0.48
typing_extensions==4.15.0
webargs==8.7.1
Werkzeug==3.1.6
